requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
APScheduler==3.11.0
asyncmy==0.2.9
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
setuptools==75.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.3
yarl==1.13.1
zope.event==5.0
zope.interface==7.0.3
